fix(gateway): close runtime transition races
